dblog-api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg2-binary==2.9.11
alembic==1.15.2
python-dotenv==1.1.0
pydantic-settings==2.9.1
httpx==0.28.1
firebase-admin==6.6.0
boto3==1.36.0
python-multipart==0.0.20
weasyprint==63.1
jinja2==3.1.5
matplotlib==3.10.1
